packet handler -> session packet handler
