+++ Skylark/Skylark/Skylark.vcxproj
@@ -141,7 +141,7 @@
     <ClInclude Include="MemoryPool.h" />
     <ClInclude Include="MPL.h" />
     <ClInclude Include="ObjectPool.h" />
-    <ClInclude Include="PacketHandler.h" />
+    <ClInclude Include="SessionPacketHandler.h" />
     <ClInclude Include="Session.h" />
     <ClInclude Include="Skylark.h" />
     <ClInclude Include="Socket.h" />
